+++ contrib/gel/vsol/vsol_box_2d.cxx
@@ -5,7 +5,7 @@
 
 vsol_box_2d::vsol_box_2d()
 {
-  // vcl_cout << "making a boxology object" << vcl_endl;
+  // vcl_cout << "making a boxology object\n";
 }
 
 vsol_box_2d::vsol_box_2d(vsol_box_2d const &b){
@@ -13,7 +13,7 @@ vsol_box_2d::vsol_box_2d(vsol_box_2d const &b){
 }
 vsol_box_2d::~vsol_box_2d()
 {
-  // vcl_cout << "deleting a boxology object" << vcl_endl;
+  // vcl_cout << "deleting a boxology object\n";
 }
 
 double vsol_box_2d::get_min_x()
@@ -78,7 +78,7 @@ bool vsol_box_2d::operator< (vsol_box_2d& b)
     this->get_max_y() <= b.get_max_y();
 }
 
-inline static bool near_same(float f1, float f2, float tolerance)
+inline static bool near_same(double f1, double f2, float tolerance)
 {
   return f1-f2<tolerance && f2-f1<tolerance;
 }
